Rename domain to Color
